feat: v2.19.0 — wire format conformance fix pass (§Phase 16 milestone 63)
Re-derives rcp/avtp.hpp's NTSCF/TSCF header bit-packing and rcp/acf.hpp's
ACF_ABB/ACF_GBB Message Info bit-packing field by field from the OPEN
Alliance TC18 Remote Control Protocol Specification v0.5.1_RC's own
bit-position diagrams, replacing the self-admitted "not full bit-for-bit
wire conformance" placeholder encoding landed at milestone 44 (v2.0.0).

- avtp.hpp: NTSCF is 12 bytes (was 14), TSCF is 24 bytes (was 18);
  sequence_num is an 8-bit rolling counter (was 16-bit); ntscf_data_length
  is an 11-bit sub-octet field; TSCF's stream_data_length lives in a
  dedicated 4-byte word after a previously-missing 4-byte reserved word.
  ByteBusId widens uint8_t -> uint16_t (byte_bus_id is an 11-bit wire
  field).
- acf.hpp: the Message Info header is 8 bytes (was 10), matching the
  specification's own two worked CRC examples that label it "8 byte
  Message Info"; acf_msg_type/acf_msg_length share a byte (7+9 bits),
  byte_bus_id is 11 bits, evt is a 4-bit field, read_size_or_segment_num is
  12 bits. Fixes two real dispatch bugs this exposed elsewhere in the tree
  (rcp/udp.hpp, rcp/record.hpp compared a raw wire byte against
  kAcfMsgTypeAbb/kAcfMsgTypeGbb directly, which almost never matches once
  acf_msg_type shares byte0 with acf_msg_length's MSB) via a new
  peek_acf_msg_type() helper. Also corrects make_response/response_kind_of:
  every response (including Acknowledge) carries rsp=1 per the
  specification's own response table; Acknowledge is distinguished by
  evt[3:0] == 0xF, not by rsp being clear.
- e2e.hpp: coverage_buffer/compute_crc/verify_crc gain a message_timestamp
  parameter and include it, at the correct byte offset, in an ACF_GBB
  message's CRC coverage (previously omitted entirely). E2eErrc::crc_error
  gains a numeric TC18 wire error code accessor (POCI_FAILURE = 12).
- acf.hpp gains ControlFlags/Message, an additive, documented thin view
  over AcfMessageInfo matching RELAY spec §15.5's canonical decoded-frame
  shape.

Closes #73, #66, #76, #89.

+++ tests/test_udp.cpp
@@ -289,7 +289,11 @@ TEST_CASE("Client::request correlates concurrent requests by byte_bus_id/transac
                            const std::vector<uint8_t>&,
                            acf::AcfMessageInfo& out_resp, std::vector<uint8_t>& out_resp_payload) {
         out_resp         = acf::make_response(req, acf::ResponseKind::ReadResponse);
-        out_resp_payload = {req.byte_bus_id, req.transaction_num};
+        // avtp::ByteBusId widened to uint16_t (rcp/avtp.hpp v2.19.0 wire
+        // conformance pass, issue cpp-RCP-04, since byte_bus_id is an
+        // 11-bit wire field) — narrow explicitly for this single-byte test
+        // payload rather than relying on an implicit narrowing conversion.
+        out_resp_payload = {static_cast<uint8_t>(req.byte_bus_id), req.transaction_num};
         return std::error_code{};
     });
